feat: added options and modifiers price with edit functionality in cart
