+++ src/apps/carts/components/index.js
@@ -3,5 +3,13 @@ import ProductOptions from './ProductOptions'
 import QuantitySelector from './QuantitySelector'
 import LoyaltyPoints from './LoyalityPoints'
 import Coupon from './Coupons'
+import EditPrice from './EditPrice'
 
-export { Modifiers, ProductOptions, QuantitySelector, LoyaltyPoints, Coupon }
+export {
+   Modifiers,
+   ProductOptions,
+   QuantitySelector,
+   LoyaltyPoints,
+   Coupon,
+   EditPrice,
+}
