+++ src/apps/carts/views/forms/ondemand/sections/Aside/CartProducts/index.jsx
@@ -1,32 +1,17 @@
 import React from 'react'
 import styled from 'styled-components'
 import { toast } from 'react-toastify'
-import { useParams } from 'react-router'
 import { useMutation, useSubscription } from '@apollo/react-hooks'
 import { LoyaltyPoints, Coupon } from '../../../../../../components'
-import {
-   Filler,
-   Flex,
-   Form,
-   IconButton,
-   Spacer,
-   Text,
-   TextButton,
-   Popup,
-   ButtonGroup,
-   ButtonTile,
-} from '@dailykit/ui'
+import { Filler, Flex, IconButton, Spacer, Text } from '@dailykit/ui'
 
 import { useManual } from '../../../state'
 import { buildImageUrl } from '../../../../../../utils'
-import { MUTATIONS, QUERIES } from '../../../../../../graphql'
+import { MUTATIONS } from '../../../../../../graphql'
 import { currencyFmt, logger } from '../../../../../../../../shared/utils'
-import {
-   CloseIcon,
-   DeleteIcon,
-   EditIcon,
-} from '../../../../../../../../shared/assets/icons'
+import { DeleteIcon } from '../../../../../../../../shared/assets/icons'
 import EmptyIllo from '../../../../../../assets/svgs/EmptyIllo'
+import EditPrice from '../../../../../../components/EditPrice'
 
 const CartProducts = () => {
    const {
@@ -37,6 +22,8 @@ const CartProducts = () => {
       tunnels,
       loyaltyPoints,
    } = useManual()
+
+   console.log('products', products)
 
    return (
       <section>
@@ -109,10 +96,6 @@ const ProductCard = ({ product, cart }) => {
       onError: () => toast.error('Failed to update the price of product.'),
    })
 
-   const [visible, setVisible] = React.useState(false)
-   const [isEdit, setIsEdit] = React.useState(false)
-   const [showPrimary, setShowPrimary] = React.useState(false)
-   const [showDanger, setShowDanger] = React.useState(false)
    const [updatedPrice, setUpdatedPrice] = React.useState({
       value: product.price,
       meta: {
@@ -121,54 +104,6 @@ const ProductCard = ({ product, cart }) => {
          isValid: true,
       },
    })
-
-   const validate = e => {
-      const { value } = e.target
-      if (value === '') {
-         setUpdatedPrice({
-            ...updatedPrice,
-            meta: {
-               errors: [],
-               isTouched: true,
-               isValid: false,
-            },
-         })
-         return
-      }
-
-      const v = parseInt(value)
-      if (isNaN(v)) {
-         setUpdatedPrice({
-            ...updatedPrice,
-            meta: {
-               errors: ['Please input numbers only!'],
-               isTouched: true,
-               isValid: false,
-            },
-         })
-         return
-      }
-      if (v <= 0) {
-         setUpdatedPrice({
-            ...updatedPrice,
-            meta: {
-               errors: ['Price should be greater than 0!'],
-               isTouched: true,
-               isValid: false,
-            },
-         })
-         return
-      }
-
-      setUpdatedPrice({
-         ...updatedPrice,
-         meta: {
-            errors: [],
-            isValid: true,
-            isTouched: true,
-         },
-      })
-   }
 
    return (
       <Styles.Card>
@@ -186,134 +121,63 @@ const ProductCard = ({ product, cart }) => {
          <Flex container alignItems="center" justifyContent="space-between">
             <Flex as="main" container flexDirection="column">
                <Text as="text2">{product.name}</Text>
-               <Flex
-                  container
-                  alignItems="center"
-                  justifyContent="space-between"
-               >
-                  {isEdit ? (
-                     <Flex
-                        container
-                        alignItems="center"
-                        justifyContent="space-between"
-                     >
-                        <Form.Group>
-                           <Form.Number
-                              value={updatedPrice.value}
-                              onChange={e =>
-                                 setUpdatedPrice({
-                                    ...updatedPrice,
-                                    value: e.target.value,
-                                 })
-                              }
-                              onBlur={validate}
-                              hasError={
-                                 updatedPrice.meta.isTouched &&
-                                 !updatedPrice.meta.isValid
-                              }
-                           />
-                           {updatedPrice.meta.isTouched &&
-                              !updatedPrice.meta.isValid &&
-                              updatedPrice.meta.errors.map((error, index) => (
-                                 <Form.Error
-                                    justifyContent="center"
-                                    key={index}
-                                 >
-                                    {error}
-                                 </Form.Error>
-                              ))}
-                        </Form.Group>
-                        <Spacer size="2px" xAxis />
-                        <IconButton
-                           type="ghost"
-                           size="sm"
-                           onClick={() => setIsEdit(!isEdit)}
-                        >
-                           <CloseIcon color="#ec3333" />
-                        </IconButton>
-                        <TextButton
-                           type="ghost"
-                           size="sm"
-                           disabled={
-                              !updatedPrice.meta.isValid || !updatedPrice.value
-                           }
-                           onClick={() => {
-                              setIsEdit(!isEdit)
-                              setShowPrimary(!showPrimary)
+               <EditPrice product={product} />
+
+               <Spacer size="2px" />
+               {product.childs.map(item => (
+                  <>
+                     <Flex alignItems="center" justifyContent="space-between">
+                        <div
+                           style={{
+                              width: '240px',
+                              whiteSpace: 'nowrap',
+                              overflow: 'hidden',
+                              textOverflow: 'ellipsis',
                            }}
                         >
-                           Update
-                        </TextButton>
+                           {item.comboProductComponent && (
+                              <Text
+                                 as="text2"
+                                 title={`${item.comboProductComponent.linkedProduct.name} (${item.comboProductComponent.label})`}
+                              >
+                                 {item.comboProductComponent.linkedProduct.name}{' '}
+                                 ({item.comboProductComponent.label})
+                              </Text>
+                           )}
+                        </div>
+                        <div>
+                           {item.customizableProductComponent && (
+                              <Text as="text2">
+                                 {item.customizableProductComponent.fullName
+                                    .split('-')[1]
+                                    .trim()}
+                              </Text>
+                           )}
+                        </div>
+                        <div style={{ marginLeft: '8px' }}>
+                           <Text as="text2" style={{ color: '#808080' }}>
+                              {item.productOption.label}
+                              <EditPrice product={item} />
+                           </Text>
+                        </div>
+
+                        <div style={{ marginLeft: '8px' }}>
+                           {item.childs.length > 0 &&
+                              item.childs.map(modi => (
+                                 <div style={{ marginLeft: '8px' }}>
+                                    <Text
+                                       as="text2"
+                                       style={{ color: '#808080' }}
+                                    >
+                                       {modi.modifierOption?.name}
+                                    </Text>
+                                    <EditPrice product={modi} />
+                                 </div>
+                              ))}
+                        </div>
                      </Flex>
-                  ) : (
-                     <Text as="text3" onMouseOver={() => setVisible(!visible)}>
-                        Price: {currencyFmt(product.price)}
-                     </Text>
-                  )}
-
-                  <Spacer xAxis size="20px" />
-                  <Popup
-                     show={showPrimary}
-                     clickOutsidePopup={() => setShowPrimary(false)}
-                  >
-                     <Popup.Actions>
-                        <Popup.Text type="primary">
-                           Closing this file will not save any changes!
-                        </Popup.Text>
-                        <Popup.Close
-                           closePopup={() => setShowPrimary(!showPrimary)}
-                        />
-                     </Popup.Actions>
-                     <Popup.ConfirmText>Are you sure?</Popup.ConfirmText>
-                     <Popup.Actions>
-                        <ButtonGroup align="left">
-                           <TextButton
-                              type="solid"
-                              onClick={() => {
-                                 setShowPrimary(!showPrimary)
-                                 if (
-                                    updatedPrice.meta.isValid &&
-                                    updatedPrice.value
-                                 ) {
-                                    update({
-                                       variables: {
-                                          id: product.id,
-                                          _set: {
-                                             unitPrice: updatedPrice.value,
-                                          },
-                                       },
-                                    })
-                                 }
-                              }}
-                           >
-                              Yes! change the price
-                           </TextButton>
-                           <TextButton
-                              type="ghost"
-                              onClick={() => setShowPrimary(!showPrimary)}
-                           >
-                              Don't want to change
-                           </TextButton>
-                        </ButtonGroup>
-                     </Popup.Actions>
-                  </Popup>
-
-                  {visible ? (
-                     <IconButton
-                        type="ghost"
-                        size="sm"
-                        onClick={() => {
-                           setIsEdit(!isEdit)
-
-                           setVisible(!visible)
-                        }}
-                     >
-                        <EditIcon />
-                     </IconButton>
-                  ) : (
-                     ''
-                  )}
-               </Flex>
+                  </>
+               ))}
             </Flex>
 
             {cart?.paymentStatus === 'PENDING' && (
